+++ esp8266_code/lib/Timer/Timer.hpp
@@ -1,0 +1,19 @@
+#pragma once
+
+#include <stdint.h>
+
+class Timer
+{
+private:
+    int64_t startTime;
+    int32_t delayInMs;
+    bool started;
+    bool finished;
+
+public:
+    Timer();
+    void Start(int32_t delayInMs);
+    void Run();
+    void Reset();
+    bool Finished() const;
+};