+++ esp8266_code/src/main.cpp
@@ -1,9 +1,48 @@
 #include <Arduino.h>
 
+#include <Led.hpp>
+#include <SerialHandler.hpp>
+
+#define BAUDRATE 9600
+#define LED_PIN D1
+#define BUTTON_1_PIN D5
+#define BUTTON_2_PIN D6
+#define SECOND 1000
+
+SerialHandler serialHandler(BAUDRATE);
+Led led(LED_PIN, SECOND);
+
 void setup() {
   // put your setup code here, to run once:
+  serialHandler.Begin();
+  led.Setup();
+  pinMode(BUTTON_1_PIN, INPUT);
+  pinMode(BUTTON_2_PIN, INPUT);
 }
 
 void loop() {
   // put your main code here, to run repeatedly:
+  serialHandler.BuildMessage();
+  led.Update();
+
+  if(serialHandler.MessageRetreived())
+  {
+    char* msg = serialHandler.GetRetreivedMessage();
+
+    if((strcmp(msg, "ACK") == 0) && (!led.IsLedOn()))
+    {
+      serialHandler.Clear();
+      led.On();
+    }
+  }
+
+  if(digitalRead(BUTTON_1_PIN) == HIGH)
+  {
+    serialHandler.SendMessage("Option1");
+  }
+
+  if(digitalRead(BUTTON_2_PIN) == HIGH)
+  {
+    serialHandler.SendMessage("Option2");
+  }
 }